moved tests from connection_test.go to tests/integration
